+++ ParselyTracker/ParselyTracker.swift
@@ -1,22 +1,11 @@
 import Foundation
+import Combine
 import UIKit
 import os.log
 
 public class Parsely {
+
     public var apikey = ""
-    var config: [String: Any] = [:]
-    private var default_config = [String: Any]()
-    private var _track: Track!
-    var track: Track {
-        return _track
-    }
-    var lastRequest: Dictionary<String, Any?>? = [:]
-    var eventQueue: EventQueue<Event> = EventQueue()
-    private var configured = false
-    private var flushTimer: Timer?
-    private var flushInterval: TimeInterval = 30
-    private var backgroundFlushTask: UIBackgroundTaskIdentifier = UIBackgroundTaskIdentifier.invalid
-    private var active: Bool = true
     public var secondsBetweenHeartbeats: TimeInterval? {
         get {
             if let secondsBtwnHeartbeats = config["secondsBetweenHeartbeats"] as! TimeInterval? {
@@ -26,16 +15,32 @@ public class Parsely {
         }
     }
     public static let sharedInstance = Parsely()
+
+    var config: [String: Any] = [:]
+    var track: Track {
+        return _track
+    }
+    var eventQueue: EventQueue<Event> = EventQueue()
+    // All tracker related state should be modified in this serial queue, to ensure thread-safty.
+    // This property is declared as internal rather than private, because we need to use it to make sure
+    // unit tests that inspect tracker's internal state can pass reliably.
+    var eventProcessor: DispatchQueue
     internal static let sharedStorage = Storage()
     lazy var visitorManager = VisitorManager()
+    internal static func getInstance() -> Parsely {
+        return Parsely()
+    }
+
+    private var _track: Track!
+    private var configured = false
+    private var flushTimer: Cancellable?
+    private var flushInterval: TimeInterval = 30
+    private var active: Bool = true
 
     private init() {
         os_log("Initializing ParselyTracker", log: OSLog.tracker, type: .info)
+        eventProcessor = DispatchQueue(label: "ly.parse.event-processor")
         _track = Track(trackerInstance: self)
-    }
-    
-    internal static func getInstance() -> Parsely {
-        return Parsely()
     }
 
     /**
@@ -66,12 +71,25 @@ public class Parsely {
      - Parameter extraData: A dictionary of additional information to send with the generated pageview event
      - Parameter siteId: The Parsely site ID for which the pageview event should be counted
      */
-    public func trackPageView(url: String,
-                              urlref: String = "",
-                              metadata: ParselyMetadata? = nil,
-                              extraData: Dictionary<String, Any>? = nil,
-                              siteId: String = "")
-    {
+    public func trackPageView(
+        url: String,
+        urlref: String = "",
+        metadata: ParselyMetadata? = nil,
+        extraData: Dictionary<String, Any>? = nil,
+        siteId: String = ""
+    ) {
+        eventProcessor.async {
+            self._trackPageView(url: url, urlref: urlref, metadata: metadata, extraData: extraData, siteId: siteId)
+        }
+    }
+
+    private func _trackPageView(
+        url: String,
+        urlref: String,
+        metadata: ParselyMetadata?,
+        extraData: Dictionary<String, Any>?,
+        siteId: String
+    ) {
         var _siteId = siteId
         if (_siteId == "") {
             _siteId = self.apikey
@@ -90,11 +108,23 @@ public class Parsely {
      - Parameter extraData: A dictionary of additional information to send with generated heartbeat events
      - Parameter siteId: The Parsely site ID for which the heartbeat events should be counted
      */
-    public func startEngagement(url: String,
-                                urlref: String = "",
-                                extraData: Dictionary<String, Any>? = nil,
-                                siteId: String = "")
-    {
+    public func startEngagement(
+        url: String,
+        urlref: String = "",
+        extraData: Dictionary<String, Any>? = nil,
+        siteId: String = ""
+    ) {
+        eventProcessor.async {
+            self._startEngagement(url: url, urlref: urlref, extraData: extraData, siteId: siteId)
+        }
+    }
+
+    private func _startEngagement(
+        url: String,
+        urlref: String,
+        extraData: Dictionary<String, Any>?,
+        siteId: String
+    ) {
         var _siteId = siteId
         if (_siteId == "") {
             _siteId = self.apikey
@@ -107,7 +137,9 @@ public class Parsely {
      previously-engaged url, after which heartbeat events will stop being sent.
      */
     public func stopEngagement() {
-        track.stopEngagement()
+        eventProcessor.async {
+            self.track.stopEngagement()
+        }
     }
 
     /**
@@ -123,14 +155,29 @@ public class Parsely {
      - Parameter extraData: A dictionary of additional information to send with generated vheartbeat events
      - Parameter siteId: The Parsely site ID for which the vheartbeat events should be counted
      */
-    public func trackPlay(url: String,
-                          urlref: String = "",
-                          videoID: String,
-                          duration: TimeInterval,
-                          metadata: ParselyMetadata? = nil,
-                          extraData: Dictionary<String, Any>? = nil,
-                          siteId: String = "")
-    {
+    public func trackPlay(
+        url: String,
+        urlref: String = "",
+        videoID: String,
+        duration: TimeInterval,
+        metadata: ParselyMetadata? = nil,
+        extraData: Dictionary<String, Any>? = nil,
+        siteId: String = ""
+    ) {
+        eventProcessor.async {
+            self._trackPlay(url: url, urlref: urlref, videoID: videoID, duration: duration, metadata: metadata, extraData: extraData, siteId: siteId)
+        }
+    }
+
+    private func _trackPlay(
+        url: String,
+        urlref: String,
+        videoID: String,
+        duration: TimeInterval,
+        metadata: ParselyMetadata?,
+        extraData: Dictionary<String, Any>?,
+        siteId: String
+    ) {
         var _siteId = siteId
         if (_siteId == "") {
             _siteId = self.apikey
@@ -143,7 +190,9 @@ public class Parsely {
      event may be sent per previously-viewed url/video combination, after which vheartbeat events will stop being sent.
      */
     public func trackPause() {
-        track.videoPause()
+        eventProcessor.async {
+            self.track.videoPause()
+        }
     }
     
     /**
@@ -154,31 +203,50 @@ public class Parsely {
      - Parameter videoId: The video ID string for the video being reset
      */
     public func resetVideo(url:String, videoID:String) {
-        track.videoReset(url: url, vId: videoID)
+        eventProcessor.async {
+            self.track.videoReset(url: url, vId: videoID)
+        }
     }
-    
+
+    /// After given `seconds`, invoke the given target-action in the event processor queue.
+    func scheduleEventProcessing(inSeconds seconds: Double, target: AnyObject, selector: Selector) -> Cancellable {
+        Just(0)
+            // From the Apple documentation it's not clear what the default tolerance is.
+            // They have an example with a .seconds(3) delay saying "3 seconds (±0.5 seconds)".
+            // To say on the safe side, we specify a 1ms tolerance, meaning "the Delay publisher may
+            // deliver elements this much sooner or later than the interval specifies."
+            .delay(for: .seconds(seconds), tolerance: .milliseconds(1), scheduler: eventProcessor)
+            .sink { _ in
+                _ = target.perform(selector)
+            }
+    }
+
     @objc private func flush() {
         if eventQueue.length() == 0 {
             return
         }
 
+        let task = BackgroundTask.begin()
+
         os_log("Flushing event queue", log: OSLog.tracker, type:.debug)
         let events = eventQueue.get()
         os_log("Got %s events", log: OSLog.tracker, type:.debug, String(describing: events.count))
         let request = RequestBuilder.buildRequest(events: events)
-        HttpClient.sendRequest(request: request!) { error in
+        HttpClient.sendRequest(request: request!, queue: eventProcessor) { error in
             if let error = error as? URLError, error.code == .notConnectedToInternet {
                 // When offline, return the events to the queue for the next flush().
                 self.eventQueue.push(contentsOf: events)
                 os_log("Network connection unavailable. Returning %s events to the queue.", String(describing: events.count))
             }
+
+            task.end()
         }
     }
     
     internal func startFlushTimer() {
         os_log("Flush timer starting", log: OSLog.tracker, type:.debug)
         if flushTimer == nil && active {
-            flushTimer = Timer.scheduledTimer(timeInterval: flushInterval, target: self, selector: #selector(flush), userInfo: nil, repeats: true)
+            flushTimer = scheduleEventProcessing(inSeconds: flushInterval, target: self, selector: #selector(flush))
             os_log("Flush timer started", log: OSLog.tracker, type:.debug)
         }
     }
@@ -186,18 +254,21 @@ public class Parsely {
     internal func pauseFlushTimer() {
         os_log("Flush timer stopping", log: OSLog.tracker, type:.debug)
         if flushTimer != nil && !active {
-            flushTimer?.invalidate()
+            flushTimer?.cancel()
             flushTimer = nil
             os_log("Flush timer stopped", log: OSLog.tracker, type:.debug)
         }
     }
 
     private func addApplicationObservers() {
-        NotificationCenter.default.addObserver(self, selector: #selector(resumeExecution(_:)), name: UIApplication.didBecomeActiveNotification, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(resumeExecution(_:)), name: UIApplication.willEnterForegroundNotification, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(suspendExecution(_:)), name: UIApplication.willResignActiveNotification, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(suspendExecution(_:)), name: UIScene.didEnterBackgroundNotification, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(suspendExecution(_:)), name: UIApplication.willTerminateNotification, object: nil)
+        let queue = OperationQueue()
+        queue.maxConcurrentOperationCount = 1
+        queue.underlyingQueue = eventProcessor
+        NotificationCenter.default.addObserver(forName: UIApplication.didBecomeActiveNotification, object: nil, queue: queue, using: self.resumeExecution(_:))
+        NotificationCenter.default.addObserver(forName: UIApplication.willEnterForegroundNotification, object: nil, queue: queue, using: self.resumeExecution(_:))
+        NotificationCenter.default.addObserver(forName: UIApplication.willResignActiveNotification, object: nil, queue: queue, using: self.suspendExecution(_:))
+        NotificationCenter.default.addObserver(forName: UIScene.didEnterBackgroundNotification, object: nil, queue: queue, using: self.suspendExecution(_:))
+        NotificationCenter.default.addObserver(forName: UIApplication.willTerminateNotification, object: nil, queue: queue, using: self.suspendExecution(_:))
     }
 
     @objc private func resumeExecution(_ notification: Notification) {
@@ -217,31 +288,45 @@ public class Parsely {
         self.active = false
         os_log("Stopping execution before background/inactive/terminate", log:OSLog.tracker, type:.info)
         hardShutdown()
-        
-        DispatchQueue.global(qos: .userInitiated).async{
-            let _self = self
-            _self.backgroundFlushTask = UIApplication.shared.beginBackgroundTask(expirationHandler:{
-                _self.endBackgroundFlushTask()
-            })
-            os_log("Flushing queue in background", log:OSLog.tracker, type:.info)
-            _self.track.sendHeartbeats()
-            _self.flush()
-            _self.endBackgroundFlushTask()
-        }
+
+        os_log("Flushing queue in background", log:OSLog.tracker, type:.info)
+        self.track.sendHeartbeats()
+        self.flush()
     }
     
     internal func hardShutdown() {
         pauseFlushTimer()
         track.pause()
     }
-    
-    private func endBackgroundFlushTask() {
-        UIApplication.shared.endBackgroundTask(backgroundFlushTask)
-        backgroundFlushTask = UIBackgroundTaskIdentifier.invalid
-    }
+
 }
 
 extension OSLog {
     private static var logger = "ParselyTracker"
     static let tracker = OSLog(subsystem: "ParselyTracker", category: "parsely_tracker")
+}
+
+private final class BackgroundTask {
+    static func begin() -> Self {
+        let task = Self()
+        task.taskIdentifier = UIApplication.shared.beginBackgroundTask {
+            task.end()
+        }
+        return task
+    }
+
+    private var taskIdentifier: UIBackgroundTaskIdentifier!
+
+    private init() {
+        // Use `begin` to create an instance.
+    }
+
+    func end() {
+        guard taskIdentifier != .invalid else {
+            return
+        }
+
+        UIApplication.shared.endBackgroundTask(taskIdentifier)
+        taskIdentifier = .invalid
+    }
 }