+++ ParselyTracker/Sampler.swift
@@ -1,4 +1,5 @@
 import Foundation
+import Combine
 import os.log
 
 let SAMPLE_RATE = TimeInterval(0.1)
@@ -32,8 +33,8 @@ class Sampler {
     var heartbeatInterval: TimeInterval
     var hasStartedSampling: Bool = false
     var accumulators: Dictionary<String, Accumulator> = [:]
-    var samplerTimer: Timer?
-    var heartbeatsTimer: Timer?
+    var samplerTimer: Cancellable?
+    var heartbeatsTimer: Cancellable?
     internal let parselyTracker: Parsely
     
     init(trackerInstance: Parsely) {
@@ -54,10 +55,12 @@ class Sampler {
     // if true, the sample() loop will accumulate time for that item.
     func sampleFn(key: String) -> Bool { return true }
 
-    public func trackKey(key: String,
-                         contentDuration: TimeInterval?,
-                         eventArgs: Dictionary<String, Any>?,
-                         resetOnExisting: Bool = false) -> Void {
+    func trackKey(
+        key: String,
+        contentDuration: TimeInterval?,
+        eventArgs: Dictionary<String, Any>?,
+        resetOnExisting: Bool = false
+    ) -> Void {
         os_log("Sampler tracked key: %s in class %@", log: OSLog.tracker, type: .debug, key, String(describing: self))
         let isNew: Bool = accumulators.index(forKey: key) == nil
         let shouldReset: Bool = !isNew && resetOnExisting
@@ -87,22 +90,22 @@ class Sampler {
     private func restartTimers() {
         os_log("Restarted Timers in %@", log: OSLog.tracker, type: .debug, String(describing: self))
         if samplerTimer != nil {
-            samplerTimer!.invalidate()
+            samplerTimer!.cancel()
         }
-        samplerTimer = Timer.scheduledTimer(timeInterval: SAMPLE_RATE, target: self, selector: #selector(sample), userInfo: nil, repeats: false)
+        samplerTimer = parselyTracker.scheduleEventProcessing(inSeconds: SAMPLE_RATE, target: self, selector: #selector(sample))
         if heartbeatsTimer != nil {
-            heartbeatsTimer!.invalidate()
+            heartbeatsTimer!.cancel()
         }
-        heartbeatsTimer = Timer.scheduledTimer(timeInterval: heartbeatInterval, target: self, selector: #selector(sendHeartbeats), userInfo: nil, repeats: false)
+        heartbeatsTimer = parselyTracker.scheduleEventProcessing(inSeconds: heartbeatInterval, target: self, selector: #selector(sendHeartbeats))
     }
 
-    public func dropKey(key: String) -> Void {
+    func dropKey(key: String) -> Void {
         os_log("Dropping Sampler key: %s", log: OSLog.tracker, type:.debug, key)
         sendHeartbeat(key: key)
         accumulators.removeValue(forKey: key)
     }
 
-    public func generateEventArgs(url: String, urlref: String, metadata: ParselyMetadata? = nil, extra_data: Dictionary<String, Any>?, idsite: String) -> Dictionary<String, Any> {
+    func generateEventArgs(url: String, urlref: String, metadata: ParselyMetadata? = nil, extra_data: Dictionary<String, Any>?, idsite: String) -> Dictionary<String, Any> {
         var eventArgs: [String: Any] = ["urlref": urlref, "url": url, "idsite": idsite]
         if (metadata != nil) {
             eventArgs["metadata"] = metadata!
@@ -127,7 +130,7 @@ class Sampler {
             trackedData.lastSampleTime = currentTime
             updateAccumulator(acc: trackedData)
         }
-        samplerTimer = Timer.scheduledTimer(withTimeInterval: SAMPLE_RATE, repeats: false) { timer in self.sample() }
+        samplerTimer = parselyTracker.scheduleEventProcessing(inSeconds: SAMPLE_RATE, target: self, selector: #selector(sample))
     }
     
     private func getHeartbeatInterval(existingTimeout: TimeInterval,
@@ -163,9 +166,7 @@ class Sampler {
         for (key, _) in accumulators {
             sendHeartbeat(key: key)
         }
-        heartbeatsTimer = Timer.scheduledTimer(withTimeInterval: TimeInterval(heartbeatInterval), repeats: false) { timer in
-            self.sendHeartbeats()
-        }
+        heartbeatsTimer = parselyTracker.scheduleEventProcessing(inSeconds: heartbeatInterval, target: self, selector: #selector(sendHeartbeats))
     }
 
     private func updateAccumulator(acc: Accumulator) -> Void {
@@ -175,11 +176,11 @@ class Sampler {
     internal func pause() {
         os_log("Paused from %@", log:OSLog.tracker, type:.debug, String(describing: self))
         if samplerTimer != nil {
-            samplerTimer!.invalidate()
+            samplerTimer!.cancel()
             samplerTimer = nil
         }
         if heartbeatsTimer != nil {
-            heartbeatsTimer!.invalidate()
+            heartbeatsTimer!.cancel()
             heartbeatsTimer = nil
         }
     }